Switched the endstop side and fixed/improved the model
Put the endstop on the other side, the original one was wrong. Adjusted
the "idler pressure" by that I mean how much the idler pushes against
the filament, had to reduce it. Added some cable rounting channels this
is just a try and proper cable management will be added in the future