Partial implementation of the calibration process
A couple of bug fixes and code polishing needed to be done.
Not yet implemented but we have the functions ready for extruder moves
to align the filament with the nozzle and then an empty menu that will
be replaced with something to edit the BOWDEN_TUBE_LENGHT varirable
defined in the config file. This will be helpful to do all the testing
and the flashing the firmware once the correct value has been found.