Easier filament tip tuning
Python script to isolate the ramming sequence and the following cooling
moves, it should help with testing and proper tweaking of the ramming sequence and evaluate the required number of cooling moves. The provided stl files have the sole purpose of creating a small Gcode file with at least one tool change. The input file is the test.gcode file and the output file that should be “printed” is the TipTest.gcode file