New Merger
this is a new merger. The choice to mount it back on the MSU was made
for simplicity and realiability. Most of the clogs happened at the
coupler level since the connection from ptfe tubing to the 3d printed tubing
can't be perfect meaning that attaching it directly to the pulley body
allows the use of a single coupler on the MSU side of things allowing
for less resistance. Furthermore, based on testing it seems that lining
the plastic tubes with PTFE doesn't necesarly improve the amount of
friction inside the merger and tends to increase it due to the added
complexity of the design and the bridging point from PTFE to regular 3D
printed material. This also helps reclaim entirely the height that we
lost from the previous merger(which on a regular 3D printer was
removing about 30% of the build volume). It will also make the overall
solution cheaper, simpler and will allow for an easier implementation of the
filament flow sensor.

The only downside will be the slightly increased tool change time but this can be negated
by increasing speed and this won't have that big of an impact on the
tool change time since the biggest time loss is when cooling of purging
the filament.